// src/unicode/generated/unicode-17.ts
export const UNICODE_VERSION = "17.0.0" as const;
export const BIDI_CLASS_VALUES = ["AL", "AN", "B", "BN", "CS", "EN", "ES", "ET", "FSI", "L", "LRE", "LRI", "LRO", "NSM", "ON", "PDF", "PDI", "R", "RLE", "RLI", "RLO", "S", "WS"] as const;
// ...
export const BIDI_DEFAULT_VALUES = ["AL", "ET", "L", "R"] as const;
export const BIDI_DEFAULT_RANGES: readonly number[] = [
  0, 1114111, 2, 1424, 1535, 3, 1536, 1983, 0, 1984, 2143, 3, 2144, 2303, 0, 8352, 8399, 1, 64285, 64335, 3, 64336, 64975, 0,
  65008, 65023, 0, 65136, 65279, 0, 67584, 68863, 3, 68864, 68927, 0, 68928, 69311, 3, 69312, 69375, 0, 69376, 69423, 3, 69424, 69487, 0,
  69488, 69631, 3, 124928, 126063, 3, 126064, 126143, 0, 126144, 126207, 3, 126208, 126287, 0, 126288, 126463, 3, 126464, 126719, 0, 126720, 126975, 3
];
export const GENERAL_CATEGORY_VALUES = ["Cc", "Cf", "Cn", "Co", "Cs", "Ll", "Lm", "Lo", "Lt", "Lu", "Mc", "Me", "Mn", "Nd", "Nl", "No", "Pc", "Pd", "Pe", "Pf", "Pi", "Po", "Ps", "Sc", "Sk", "Sm", "So", "Zl", "Zp", "Zs"] as const;
// ...
export const GRAPHEME_BREAK_VALUES = ["CR", "Control", "Extend", "L", "LF", "LV", "LVT", "Prepend", "Regional_Indicator", "SpacingMark", "T", "V", "ZWJ"] as const;
// ...
export const INDIC_CONJUNCT_BREAK_VALUES = ["Consonant", "Extend", "Linker"] as const;
// ...
export const LINE_BREAK_VALUES = ["AI", "AK", "AL", "AP", "AS", "B2", "BA", "BB", "BK", "CB", "CJ", "CL", "CM", "CP", "CR", "EB", "EM", "EX", "GL", "H2", "H3", "HH", "HL", "HY", "ID", "IN", "IS", "JL", "JT", "JV", "LF", "NL", "NS", "NU", "OP", "PO", "PR", "QU", "RI", "SA", "SG", "SP", "SY", "VF", "VI", "WJ", "XX", "ZW", "ZWJ"] as const;
// ...
export const EAST_ASIAN_WIDTH_VALUES = ["A", "F", "H", "N", "Na", "W"] as const;
// ...
export const BIDI_BRACKET_MAP: readonly number[] = [
  40, 41, 1, 41, 40, 2, 91, 93, 1, 93, 91, 2, 123, 125, 1, 125, 123, 2, 3898, 3899, 1, 3899, 3898, 2,
  3900, 3901, 1, 3901, 3900, 2, 5787, 5788, 1, 5788, 5787, 2, 8261, 8262, 1, 8262, 8261, 2, 8317, 8318, 1, 8318, 8317, 2,
  8333, 8334, 1, 8334, 8333, 2, 8968, 8969, 1, 8969, 8968, 2, 8970, 8971, 1, 8971, 8970, 2, 9001, 9002, 1, 9002, 9001, 2,
  10088, 10089, 1, 10089, 10088, 2, 10090, 10091, 1, 10091, 10090, 2, 10092, 10093, 1, 10093, 10092, 2, 10094, 10095, 1, 10095, 10094, 2,
  10096, 10097, 1, 10097, 10096, 2, 10098, 10099, 1, 10099, 10098, 2, 10100, 10101, 1, 10101, 10100, 2, 10181, 10182, 1, 10182, 10181, 2,
  10214, 10215, 1, 10215, 10214, 2, 10216, 10217, 1, 10217, 10216, 2, 10218, 10219, 1, 10219, 10218, 2, 10220, 10221, 1, 10221, 10220, 2,
  10222, 10223, 1, 10223, 10222, 2, 10627, 10628, 1, 10628, 10627, 2, 10629, 10630, 1, 10630, 10629, 2, 10631, 10632, 1, 10632, 10631, 2,
  10633, 10634, 1, 10634, 10633, 2, 10635, 10636, 1, 10636, 10635, 2, 10637, 10640, 1, 10638, 10639, 2, 10639, 10638, 1, 10640, 10637, 2,
  10641, 10642, 1, 10642, 10641, 2, 10643, 10644, 1, 10644, 10643, 2, 10645, 10646, 1, 10646, 10645, 2, 10647, 10648, 1, 10648, 10647, 2,
  10712, 10713, 1, 10713, 10712, 2, 10714, 10715, 1, 10715, 10714, 2, 10748, 10749, 1, 10749, 10748, 2, 11810, 11811, 1, 11811, 11810, 2,
  11812, 11813, 1, 11813, 11812, 2, 11814, 11815, 1, 11815, 11814, 2, 11816, 11817, 1, 11817, 11816, 2, 11861, 11862, 1, 11862, 11861, 2,
  11863, 11864, 1, 11864, 11863, 2, 11865, 11866, 1, 11866, 11865, 2, 11867, 11868, 1, 11868, 11867, 2, 12296, 12297, 1, 12297, 12296, 2,
  12298, 12299, 1, 12299, 12298, 2, 12300, 12301, 1, 12301, 12300, 2, 12302, 12303, 1, 12303, 12302, 2, 12304, 12305, 1, 12305, 12304, 2,
  12308, 12309, 1, 12309, 12308, 2, 12310, 12311, 1, 12311, 12310, 2, 12312, 12313, 1, 12313, 12312, 2, 12314, 12315, 1, 12315, 12314, 2,
  65113, 65114, 1, 65114, 65113, 2, 65115, 65116, 1, 65116, 65115, 2, 65117, 65118, 1, 65118, 65117, 2, 65288, 65289, 1, 65289, 65288, 2,
  65339, 65341, 1, 65341, 65339, 2, 65371, 65373, 1, 65373, 65371, 2, 65375, 65376, 1, 65376, 65375, 2, 65378, 65379, 1, 65379, 65378, 2
];
export const BIDI_BRACKET_CANONICAL_MAP: readonly number[] = [
  9001, 12296, 9002, 12297
];